+++ lua/highlights.lua
@@ -2,7 +2,7 @@ local colors = require("settings").colors
 
 local base00 = "#282c34"
 local base01 = "#353b45"
-local base02 = "#3e4451"
+-- local base02 = "#3e4451"
 local base03 = "#545862"
 local base04 = "#565c64"
 local base05 = "#abb2bf"
@@ -17,285 +17,263 @@ local base0D = colors.blue
 local base0E = colors.purple
 local base0F = colors.tomato
 
-local cmd = vim.cmd
-
-local function highlight(all_highlights)
-    for group, ui in pairs(all_highlights) do
-        local guifg = ui.guifg or "NONE"
-        local guibg = ui.guibg or "NONE"
-        local gui = ui.gui or "NONE"
-        if gui == "underline" or gui == "undercurl" then
-            local guisp = ui.guisp or "NONE"
-
-            cmd(string.format("hi %s guifg=%s guibg=%s gui=%s guisp=%s", group, guifg, guibg, gui, guisp))
-        else
-            cmd(string.format("hi %s guifg=%s guibg=%s gui=%s", group, guifg, guibg, gui))
-        end
-    end
+local function nvim_set_hl(name, val)
+    vim.api.nvim_set_hl(0, name, val)
 end
 
--- Credit https://github.com/RRethy/nvim-base16/blob/master/lua/base16-colorscheme.lua
-highlight({
-    -- Conceal = { guifg = base0D, guibg = base00 },
-    Cursor = { guifg = base00 },
-    CursorLine = { guibg = colors.cursorLine },
-    CursorLineNr = { guifg = base0A, guibg = colors.cursorLine },
-    Directory = { guifg = base0D },
-    EndOfBuffer = { guifg = colors.comment },
-    ErrorMsg = { guifg = base08, guibg = colors.background },
-    Folded = { guifg = colors.comment, guibg = colors.cursorLine },
-    FoldColumn = { guifg = colors.comment, guibg = colors.background },
-    IncSearch = { guifg = base01, guibg = base09 },
-    LineNr = { guifg = colors.lineNumber, guibg = colors.background },
-    -- Macro = { guifg = base08 },
-    MatchParen = { guibg = base03 },
-    MoreMsg = { guifg = base0B },
-    NonText = { guifg = colors.comment },
-    Normal = { guibg = colors.background },
-    NormalFloat = { guifg = base05, guibg = colors.dark_background },
-    NormalNC = { guibg = colors.background },
-    FloatBorder = { guifg = base05, guibg = base00 },
-    ColorColumn = { guibg = colors.cursorLine },
+nvim_set_hl("Cursor", { fg = base00 })
+nvim_set_hl("CursorLine", { bg = colors.cursorLine })
+nvim_set_hl("CursorLineNr", { fg = base0A, bg = colors.cursorLine })
+nvim_set_hl("Directory", { fg = base0D })
+nvim_set_hl("EndOfBuffer", { fg = colors.comment })
+nvim_set_hl("ErrorMsg", { fg = base08, bg = colors.background })
+nvim_set_hl("Folded", { fg = colors.comment, bg = colors.cursorLine })
+nvim_set_hl("FoldColumn", { fg = colors.comment, bg = colors.background })
+nvim_set_hl("IncSearch", { fg = base01, bg = base09 })
+nvim_set_hl("LineNr", { fg = colors.lineNumber, bg = colors.background })
+nvim_set_hl("MatchParen", { bg = base03 })
+nvim_set_hl("MoreMsg", { fg = base0B })
+nvim_set_hl("NonText", { fg = colors.comment })
+nvim_set_hl("Normal", { bg = colors.background })
+nvim_set_hl("NormalFloat", { fg = base05, bg = colors.dark_background })
+nvim_set_hl("NormalNC", { bg = colors.background })
+nvim_set_hl("FloatBorder", { fg = base05, bg = base00 })
+nvim_set_hl("ColorColumn", { bg = colors.cursorLine })
 
-    Pmenu = { guifg = base05, guibg = colors.dark_background },
-    PmenuSbar = { guifg = base05, guibg = colors.dark_background },
-    PmenuSel = { guifg = colors.dark_background, guibg = base0B },
-    PmenuThumb = { guifg = base05, guibg = colors.scrollbar },
+nvim_set_hl("Pmenu", { fg = base05, bg = colors.dark_background })
+nvim_set_hl("PmenuSbar", { fg = base05, bg = colors.dark_background })
+nvim_set_hl("PmenuSel", { fg = colors.dark_background, bg = base0B })
+nvim_set_hl("PmenuThumb", { fg = base05, bg = colors.scrollbar })
 
-    CmpItemMenu = { guifg = base05, guibg = colors.dark_background },
-    CmpItemKind = { guifg = colors.cyan },
-    CmpItemKindMethod = { guifg = base0A },
-    CmpItemKindFunction = { guifg = base0E },
-    CmpItemAbbrMatchFuzzy = { guifg = base0B },
-    CmpItemAbbrMatch = { guifg = base0B },
-    CmpItemAbbr = { guifg = base05 },
-    CmpItemAbbrDeprecated = { guifg = "#ffffff" },
+nvim_set_hl("CmpItemMenu", { fg = base05, bg = colors.dark_background })
+nvim_set_hl("CmpItemKind", { fg = base0C })
+nvim_set_hl("CmpItemKindMethod", { fg = base0A })
+nvim_set_hl("CmpItemKindFunction", { fg = base0E })
+nvim_set_hl("CmpItemAbbrMatchFuzzy", { fg = base0B })
+nvim_set_hl("CmpItemAbbrMatch", { fg = base0B })
+nvim_set_hl("CmpItemAbbr", { fg = base05 })
+nvim_set_hl("CmpItemAbbrDeprecated", { fg = "#ffffff" })
 
-    Question = { guifg = base0D },
-    -- QuickFixLine = { guibg = base01 },
-    Search = { guifg = base01, guibg = base0A },
-    SignColumn = { guifg = base04, guibg = colors.background },
-    SpecialKey = { guifg = base03 },
-    StatusLine = { guifg = colors.dark_background, guibg = colors.dark_background },
-    StatucLineNC = { guifg = colors.dark_background, guibg = colors.dark_background, gui = "underline" },
-    Substitute = { guifg = base01, guibg = base0A },
-    TermCursor = { guifg = colors.background, guibg = base0C },
-    TermCursorNC = { guifg = colors.background, guibg = base05 },
-    Title = { guifg = base0D },
-    VertSplit = { guifg = colors.dark_background, guibg = colors.dark_background },
-    Visual = { guibg = colors.cursorLine },
-    VisualNOS = { guifg = base08 },
-    WarningMsg = { guifg = base08 },
-    -- WildMenu = { guifg = base00, guibg = base05 },
+nvim_set_hl("Question", { fg = base0D })
+nvim_set_hl("Search", { fg = base01, bg = base0A })
+nvim_set_hl("SignColumn", { fg = base04, bg = colors.background })
+nvim_set_hl("SpecialKey", { fg = base03 })
+nvim_set_hl("StatusLine", { fg = colors.dark_background, bg = colors.dark_background })
+nvim_set_hl("StatucLineNC", { fg = colors.dark_background, bg = colors.dark_background, underline = true })
+nvim_set_hl("Substitute", { fg = base01, bg = base0A })
+nvim_set_hl("TermCursor", { fg = colors.background, bg = base0C })
+nvim_set_hl("TermCursorNC", { fg = colors.background, bg = base05 })
+nvim_set_hl("Title", { fg = base0D })
+nvim_set_hl("VertSplit", { fg = colors.dark_background, bg = colors.dark_background })
+nvim_set_hl("Visual", { bg = colors.cursorLine })
+nvim_set_hl("VisualNOS", { fg = base08 })
+nvim_set_hl("WarningMsg", { fg = base08 })
 
-    -- Spell
-    SpellBad = { gui = "undercurl", guisp = base08 },
-    SpellLocal = { gui = "undercurl", guisp = base0C },
-    SpellCap = { gui = "undercurl", guisp = base0D },
-    SpellRare = { gui = "undercurl", guisp = base0E },
+nvim_set_hl("SpellBad", { undercurl = true, sp = base08 })
+nvim_set_hl("SpellLocal", { undercurl = true, sp = base0C })
+nvim_set_hl("SpellCap", { undercurl = true, sp = base0D })
+nvim_set_hl("SpellRare", { undercurl = true, sp = base0E })
 
-    -- Syntax
-    Comment = { guifg = base03, gui = "italic" },
-    Constant = { guifg = base09, gui = "italic" },
-    String = { guifg = base0B },
-    Character = { guifg = base0C },
-    Number = { guifg = base09, gui = "italic" },
-    Boolean = { guifg = base09, gui = "italic" },
-    Float = { guifg = base09, gui = "italic" },
-    Identifier = { guifg = base08 },
-    Fuction = { guifg = base0D },
-    Statement = { guifg = base0E },
-    Conditional = { guifg = base0E, gui = "italic" },
-    Repeat = { guifg = base0E, gui = "italic" },
-    Label = { guifg = base0E },
-    Operator = { guifg = base05 },
-    Keyword = { guifg = base0E, gui = "italic" },
-    Exception = { guifg = base0E },
-    PreProc = { guifg = base0A },
-    Include = { guifg = base0D },
-    Define = { guifg = base0D },
-    Marco = { guifg = base0D },
-    PreCondit = { guifg = base0D },
-    Type = { guifg = base0D },
-    TypeDef = { guifg = base0A },
-    StorageClass = { guifg = base0E },
-    Structure = { guifg = base0E },
-    Special = { guifg = base0C },
-    SpecialChar = { guifg = base0C },
-    Tag = { guifg = base0A, gui = "underline" },
-    Delimiter = { guifg = base0F },
-    SpecialComment = { guifg = base0C },
-    Debug = { guifg = base08 },
-    Underlined = { guifg = base05, gui = "bold,underline" },
-    Error = { guifg = base00, guibg = base08 },
-    Todo = { guifg = base0A, guibg = base01 },
+nvim_set_hl("Comment", { fg = base03, italic = true })
+nvim_set_hl("Constant", { fg = base09, italic = true })
+nvim_set_hl("String", { fg = base0B })
+nvim_set_hl("Character", { fg = base0C })
+nvim_set_hl("Number", { fg = base09, italic = true })
+nvim_set_hl("Boolean", { fg = base09, italic = true })
+nvim_set_hl("Float", { fg = base09, italic = true })
+nvim_set_hl("Identifier", { fg = base08 })
+nvim_set_hl("Fuction", { fg = base0D })
+nvim_set_hl("Statement", { fg = base0E })
+nvim_set_hl("Conditional", { fg = base0E, italic = true })
+nvim_set_hl("Repeat", { fg = base0E, italic = true })
+nvim_set_hl("Label", { fg = base0E })
+nvim_set_hl("Operator", { fg = base05 })
+nvim_set_hl("Keyword", { fg = base0E, italic = true })
+nvim_set_hl("Exception", { fg = base0E })
+nvim_set_hl("PreProc", { fg = base0A })
+nvim_set_hl("Include", { fg = base0D })
+nvim_set_hl("Define", { fg = base0D })
+nvim_set_hl("Marco", { fg = base0D })
+nvim_set_hl("PreCondit", { fg = base0D })
+nvim_set_hl("Type", { fg = base0D })
+nvim_set_hl("TypeDef", { fg = base0A })
+nvim_set_hl("StorageClass", { fg = base0E })
+nvim_set_hl("Structure", { fg = base0E })
+nvim_set_hl("Special", { fg = base0C })
+nvim_set_hl("SpecialChar", { fg = base0C })
+nvim_set_hl("Tag", { fg = base0A, underline = true })
+nvim_set_hl("Delimiter", { fg = base0F })
+nvim_set_hl("SpecialComment", { fg = base0C })
+nvim_set_hl("Debug", { fg = base08 })
+nvim_set_hl("Underlined", { fg = base05, bold = true, underline = true })
+nvim_set_hl("Error", { fg = base00, bg = base08 })
+nvim_set_hl("Todo", { fg = base0A, bg = base01 })
 
-    -- LSP
-    LspReferenceText = { gui = "underline", guisp = base04 },
-    LspReferenceRead = { gui = "underline", guisp = base04 },
-    LspReferenceWrite = { gui = "underline", guisp = base04 },
-    DiagnosticError = { guifg = base08 },
-    DiagnosticWarn = { guifg = base0A },
-    DiagnosticInfo = { guifg = base0D },
-    DiagnosticHint = { guifg = base0B },
-    DiagnosticUnderlineError = { gui = "undercurl", guisp = base08 },
-    DiagnosticUnderlineWarn = { gui = "undercurl", guisp = base0A },
-    DiagnosticUnderlineInfo = { gui = "undercurl", guisp = base0D },
-    DiagnosticUnderlineHint = { gui = "undercurl", guisp = base0B },
+nvim_set_hl("LspReferenceText", { underline = true, sp = base04 })
+nvim_set_hl("LspReferenceRead", { underline = true, sp = base04 })
+nvim_set_hl("LspReferenceWrite", { underline = true, sp = base04 })
+nvim_set_hl("DiagnosticError", { fg = base08 })
+nvim_set_hl("DiagnosticWarn", { fg = base0A })
+nvim_set_hl("DiagnosticInfo", { fg = base0D })
+nvim_set_hl("DiagnosticHint", { fg = base0B })
+nvim_set_hl("DiagnosticUnderlineError", { undercurl = true, sp = base08 })
+nvim_set_hl("DiagnosticUnderlineWarn", { undercurl = true, sp = base0A })
+nvim_set_hl("DiagnosticUnderlineInfo", { undercurl = true, sp = base0D })
+nvim_set_hl("DiagnosticUnderlineHint", { undercurl = true, sp = base0B })
 
-    CodeActionSign = { guifg = base0A },
+nvim_set_hl("CodeActionSign", { fg = base0A })
 
-    AerialNormal = { guibg = colors.dark_background },
-    DapUINormal = { guibg = colors.dark_background },
-    DiffviewFilesNormal = { guibg = colors.dark_background },
+nvim_set_hl("AerialNormal", { bg = colors.dark_background })
+nvim_set_hl("DapUINormal", { bg = colors.dark_background })
+nvim_set_hl("DiffviewFilesNormal", { bg = colors.dark_background })
 
-    -- Dap-ui
-    -- DapUIVariable = { guibg = colors.dark_background },
-    -- DapUIValue = { guibg = colors.dark_background },
-    -- DapUIFrameName = { guibg = colors.dark_background },
+nvim_set_hl("TSAnnotation", { fg = base0A })
+nvim_set_hl("TSAttribute", { fg = base0A })
+nvim_set_hl("TSBoolean", { fg = base09, italic = true })
+nvim_set_hl("TSCharacter", { fg = base0C })
+nvim_set_hl("TSComment", { fg = base03, italic = true })
+nvim_set_hl("TSConstructor", { fg = base0C })
+nvim_set_hl("TSConditional", { fg = base0E, italic = true })
+nvim_set_hl("TSConstant", { fg = base09, italic = true })
+nvim_set_hl("TSConstBuiltin", { fg = base0E })
+nvim_set_hl("TSConstMacro", { fg = base0E })
+nvim_set_hl("TSError", { fg = base08 })
+nvim_set_hl("TSException", { fg = base0E })
+nvim_set_hl("TSField", { fg = base08 })
+nvim_set_hl("TSFloat", { fg = base09, italic = true })
+nvim_set_hl("TSFunction", { fg = base0D })
+nvim_set_hl("TSFuncBuiltin", { fg = base0E, italic = true })
+nvim_set_hl("TSFuncMarcro", { fg = base0D })
+nvim_set_hl("TSInclude", { fg = base0D })
+nvim_set_hl("TSKeyword", { fg = base0E, italic = true })
+nvim_set_hl("TSKeywordFunction", { fg = base0E, italic = true })
+nvim_set_hl("TSKeywordOperator", { fg = base0E, italic = true })
+nvim_set_hl("TSLabel", { fg = base0E })
+nvim_set_hl("TSMethod", { fg = base0D })
+nvim_set_hl("TSNamespace", { fg = base0E })
+nvim_set_hl("TSNone", { fg = base0E })
+nvim_set_hl("TSNumber", { fg = base09, italic = true })
+nvim_set_hl("TSOperator", { fg = base0E })
+nvim_set_hl("TSParameter", { fg = base08 })
+nvim_set_hl("TSParameterReference", { fg = base08 })
+nvim_set_hl("TSProperty", { fg = base0A })
+nvim_set_hl("TSPunctDelimiter", { fg = base0F })
+nvim_set_hl("TSPunctBracket", { fg = base0C })
+nvim_set_hl("TSPunctSpecial", { fg = base0F })
+nvim_set_hl("TSRepeat", { fg = base0E, italic = true })
+nvim_set_hl("TSString", { fg = base0B })
+nvim_set_hl("TSStringRegex", { fg = base0B })
+nvim_set_hl("TSStringEscape", { fg = base0C })
+nvim_set_hl("TSStringSpecial", { fg = base0B })
+nvim_set_hl("TSSymbol", { fg = base0B })
+nvim_set_hl("TSTag", { fg = base0A })
+nvim_set_hl("TSTagAttribute", { fg = base0A })
+nvim_set_hl("TSTagDelimiter", { fg = base0F })
+nvim_set_hl("TSText", { fg = base08 })
+nvim_set_hl("TSStrong", { bold = true })
+nvim_set_hl("TSEmphasis", { fg = base09, italic = true })
+nvim_set_hl("TSUnderline", { fg = base00, underline = true })
+nvim_set_hl("TSStrike", { fg = base00, strikethrough = true })
+nvim_set_hl("TSTitle", { fg = base0D })
+nvim_set_hl("TSLiteral", { fg = base09 })
+nvim_set_hl("TSURI", { fg = base09, underline = true })
+nvim_set_hl("TSMath", { fg = base0E })
+nvim_set_hl("TSType", { fg = base0A, italic = true })
+nvim_set_hl("TSTypeBuiltin", { fg = base0E })
+nvim_set_hl("TSVariable", { fg = base08 })
+nvim_set_hl("TSVariableBuiltin", { fg = base0E, italic = true })
 
-    -- Treesitter
-    TSAnnotation = { guifg = base0A },
-    TSAttribute = { guifg = base0A },
-    TSBoolean = { guifg = base09, gui = "italic" },
-    TSCharacter = { guifg = base0C },
-    TSComment = { guifg = base03, gui = "italic" },
-    TSConstructor = { guifg = base0C },
-    TSConditional = { guifg = base0E, gui = "italic" },
-    TSConstant = { guifg = base09, gui = "italic" },
-    TSConstBuiltin = { guifg = base0E },
-    TSConstMacro = { guifg = base0E },
-    TSError = { guifg = base08 },
-    TSException = { guifg = base0E },
-    TSField = { guifg = base08 },
-    TSFloat = { guifg = base09, gui = "italic" },
-    TSFunction = { guifg = base0D },
-    TSFuncBuiltin = { guifg = base0E, gui = "italic" },
-    TSFuncMarcro = { guifg = base0D },
-    TSInclude = { guifg = base0D },
-    TSKeyword = { guifg = base0E, gui = "italic" },
-    TSKeywordFunction = { guifg = base0E, gui = "italic" },
-    TSKeywordOperator = { guifg = base0E, gui = "italic" },
-    TSLabel = { guifg = base0E },
-    TSMethod = { guifg = base0D },
-    TSNamespace = { guifg = base0E },
-    TSNone = { guifg = base0E },
-    TSNumber = { guifg = base09, gui = "italic" },
-    TSOperator = { guifg = base0E },
-    TSParameter = { guifg = base08 },
-    TSParameterReference = { guifg = base08 },
-    TSProperty = { guifg = base0A },
-    TSPunctDelimiter = { guifg = base0F },
-    TSPunctBracket = { guifg = base0C },
-    TSPunctSpecial = { guifg = base0F },
-    TSRepeat = { guifg = base0E, gui = "italic" },
-    TSString = { guifg = base0B },
-    TSStringRegex = { guifg = base0B },
-    TSStringEscape = { guifg = base0C },
-    TSStringSpecial = { guifg = base0B },
-    TSSymbol = { guifg = base0B },
-    TSTag = { guifg = base0A },
-    TSTagAttribute = { guifg = base0A },
-    TSTagDelimiter = { guifg = base0F },
-    TSText = { guifg = base08 },
-    TSStrong = { gui = "bold" },
-    TSEmphasis = { guifg = base09, gui = "italic" },
-    TSUnderline = { guifg = base00, gui = "underline" },
-    TSStrike = { guifg = base00, gui = "strikethrough" },
-    TSTitle = { guifg = base0D },
-    TSLiteral = { guifg = base09 },
-    TSURI = { guifg = base09, gui = "underline" },
-    TSMath = { guifg = base0E },
-    TSType = { guifg = base0A, gui = "italic" },
-    TSTypeBuiltin = { guifg = base0E },
-    TSVariable = { guifg = base08 },
-    TSVariableBuiltin = { guifg = base0E, "italic" },
+nvim_set_hl("DiffAdd", { fg = base0B, bg = base00 })
+nvim_set_hl("DiffChange", { fg = base03, bg = base00 })
+nvim_set_hl("DiffDelete", { fg = base08, bg = base00 })
+nvim_set_hl("DiffText", { fg = base0D, bg = base00 })
+nvim_set_hl("DiffAdded", { fg = base0B, bg = base00 })
+nvim_set_hl("DiffFile", { fg = base08, bg = base00 })
+nvim_set_hl("DiffNewFile", { fg = base0B, bg = base00 })
+nvim_set_hl("DiffLine", { fg = base0D, bg = base00 })
+nvim_set_hl("DiffRemoved", { fg = base08, bg = base00 })
 
-    -- Diff
-    DiffAdd = { guifg = base0B, guibg = base00 },
-    DiffChange = { guifg = base03, guibg = base00 },
-    DiffDelete = { guifg = base08, guibg = base00 },
-    DiffText = { guifg = base0D, guibg = base00 },
-    DiffAdded = { guifg = base0B, guibg = base00 },
-    DiffFile = { guifg = base08, guibg = base00 },
-    DiffNewFile = { guifg = base0B, guibg = base00 },
-    DiffLine = { guifg = base0D, guibg = base00 },
-    DiffRemoved = { guifg = base08, guibg = base00 },
+nvim_set_hl("IndentBlanklineContextChar", { fg = base0A })
 
-    IndentBlanklineContextChar = { guifg = base0A },
+nvim_set_hl("StatuslineBackground", { bg = colors.dark_background })
+nvim_set_hl("StatuslineEmptyBackground", { fg = colors.dark_background, bg = colors.dark_background })
+nvim_set_hl("StatuslineDiagnosticsError", { fg = base08, bg = colors.dark_background })
+nvim_set_hl("StatuslineDiagnosticsWarning", { fg = base0A, bg = colors.dark_background })
+nvim_set_hl("StatuslineDiffAdded", { fg = base0B, bg = colors.dark_background })
+nvim_set_hl("StatuslineDiffModified", { fg = base0D, bg = colors.dark_background })
+nvim_set_hl("StatuslineDiffRemoved", { fg = base08, bg = colors.dark_background })
+nvim_set_hl("StatuslineSmiley", { fg = base0A, bg = colors.dark_background })
+nvim_set_hl("StatuslineTest", { fg = base0A, bg = colors.dark_background })
+nvim_set_hl("StatuslinePassed", { fg = base0B, bg = colors.dark_background })
+nvim_set_hl("StatuslineFailed", { fg = base08, bg = colors.dark_background })
+nvim_set_hl("StatuslineRunning", { fg = base0D, bg = colors.dark_background })
 
-    -- Status line
-    StatuslineBackground = { guibg = colors.dark_background },
-    StatuslineEmptyBackground = { guifg = colors.dark_background, guibg = colors.dark_background },
-    StatuslineDiagnosticsError = { guifg = base08, guibg = colors.dark_background },
-    StatuslineDiagnosticsWarning = { guifg = base0A, guibg = colors.dark_background },
-    StatuslineDiffAdded = { guifg = base0B, guibg = colors.dark_background },
-    StatuslineDiffModified = { guifg = base0D, guibg = colors.dark_background },
-    StatuslineDiffRemoved = { guifg = base08, guibg = colors.dark_background },
-    StatuslineSmiley = { guifg = base0A, guibg = colors.dark_background },
-    StatuslineTest = { guifg = base0A, guibg = colors.dark_background },
-    StatuslinePassed = { guifg = base0B, guibg = colors.dark_background },
-    StatuslineFailed = { guifg = base08, guibg = colors.dark_background },
-    StatuslineRunning = { guifg = base0D, guibg = colors.dark_background },
+nvim_set_hl("DapBreakpoint", { fg = base08 })
+nvim_set_hl("DapStopped", { fg = base0B })
 
-    DapBreakpoint = { guifg = base08 },
-    DapStopped = { guifg = base0B },
+nvim_set_hl("GitSignsAdd", { fg = base0B, bg = colors.background })
+nvim_set_hl("GitSignsChange", { fg = base0D, bg = colors.background })
+nvim_set_hl("GitSignsDelete", { fg = base08, bg = colors.background })
+nvim_set_hl("GitSignsCurrentLineBlame", { fg = colors.comment })
 
-    -- GitSign
-    GitSignsAdd = { guifg = base0B, guibg = colors.background },
-    GitSignsChange = { guifg = base0D, guibg = colors.background },
-    GitSignsDelete = { guifg = base08, guibg = colors.background },
-    GitSignsCurrentLineBlame = { guifg = colors.comment },
+nvim_set_hl("TroubleNormal", { bg = colors.dark_background })
 
-    TroubleNormal = { guibg = colors.dark_background },
+nvim_set_hl("YanilNormal", { bg = colors.dark_background })
+nvim_set_hl("YanilDeviconDirectory", { fg = "#cccccc" })
+nvim_set_hl("YanilDeviconDirectory", { fg = "#cccccc" })
+nvim_set_hl("YanilDeviconFile", { fg = "#cccccc" })
+nvim_set_hl("YanilTreeRoot", { fg = colors.white })
+nvim_set_hl("YanilGitModified", { fg = base0A })
+nvim_set_hl("YanilGitDirty", { fg = base0A })
+nvim_set_hl("YanilGitClean", { fg = base0B })
+nvim_set_hl("YanilIndentMarker", { fg = colors.indent })
 
-    YanilNormal = { guibg = colors.dark_background },
-    YanilDeviconDirectory = { guifg = "#cccccc" },
-    YanilDeviconFile = { guifg = "#cccccc" },
-    YanilTreeRoot = { guifg = colors.white },
-    YanilGitModified = { guifg = base0A },
-    YanilGitDirty = { guifg = base0A },
-    YanilGitClean = { guifg = base0B },
-    YanilIndentMarker = { guifg = colors.indent },
+nvim_set_hl("TelescopeBorder", { fg = colors.border })
+nvim_set_hl("TelescopePromptBorder", { fg = colors.border })
+nvim_set_hl("TelescopeResultsBorder", { fg = colors.border })
+nvim_set_hl("TelescopePreviewBorder", { fg = colors.border })
 
-    -- Nvim-tree
-    -- NvimTreeFolderIcon = { guifg = base0D },
-    -- NvimTreeFolderName = { guifg = colors.white },
-    -- NvimTreeOpenedFolderName = { guifg = colors.white },
-    -- NvimTreeEmptyFolderName = { guifg = colors.white },
-    -- NvimTreeOpenedFile = { guifg = colors.white },
-    -- NvimTreeIndentMarker = { guifg = colors.indent },
-    -- NvimTreeVertSplit = { guifg = colors.dark_background, guibg = colors.dark_background },
-    -- NvimTreeStatuslineNc = { guifg = colors.dark_background, guibg = colors.dark_background, gui = 'underline' },
-    -- NvimTreeNormal = { guibg = colors.dark_background },
-    -- NvimTreeRootFolder = { guifg = colors.white, gui = 'bold' },
+nvim_set_hl("DapUIFloatBorder", { fg = colors.border })
 
-    -- Telescope
-    TelescopeBorder = { guifg = colors.border },
-    TelescopePromptBorder = { guifg = colors.border },
-    TelescopeResultsBorder = { guifg = colors.border },
-    TelescopePreviewBorder = { guifg = colors.border },
+nvim_set_hl("DashboardHeader", { fg = colors.comment })
 
-    DapUIFloatBorder = { guifg = colors.border },
+nvim_set_hl("ScrollView", { bg = colors.scrollbar })
 
-    DashboardHeader = { guifg = colors.comment },
+nvim_set_hl("ScrollbarErrorHandle", { fg = base08 })
+nvim_set_hl("ScrollbarWarnHandle", { fg = base0A })
+nvim_set_hl("ScrollbarInfoHandle", { fg = base0D })
+nvim_set_hl("ScrollbarHintHandle", { fg = base0B })
+nvim_set_hl("ScrollbarError", { fg = base08 })
+nvim_set_hl("ScrollbarWarn", { fg = base0A })
+nvim_set_hl("ScrollbarInfo", { fg = base0D })
+nvim_set_hl("ScrollbarHint", { fg = base0B })
 
-    ScrollView = { guibg = colors.scrollbar },
+nvim_set_hl("UfoPreviewThumb", { bg = colors.background, fg = base08 })
+nvim_set_hl("UfoPreviewSbar", { bg = colors.background, fg = base08 })
+nvim_set_hl("UfoFoldedEllipsis", { bg = colors.background, fg = base08 })
 
-    ScrollbarErrorHandle = { guifg = base08 },
-    ScrollbarWarnHandle = { guifg = base0A },
-    ScrollbarInfoHandle = { guifg = base0D },
-    ScrollbarHintHandle = { guifg = base0B },
-    ScrollbarError = { guifg = base08 },
-    ScrollbarWarn = { guifg = base0A },
-    ScrollbarInfo = { guifg = base0D },
-    ScrollbarHint = { guifg = base0B },
+-- Conceal = { guifg = base0D, guibg = base00 },
+-- Macro = { guifg = base08 },
 
-    -- UFO
-    UfoPreviewThumb = { guibg = colors.background, guifg = base08 },
-    UfoPreviewSbar = { guibg = colors.background, guifg = base08 },
-    UfoFoldedEllipsis = { guibg = colors.background, guifg = base08 },
-})
+-- QuickFixLine = { guibg = base01 },
+-- WildMenu = { guifg = base00, guibg = base05 },
+
+-- Dap-ui
+-- DapUIVariable = { guibg = colors.dark_background },
+-- DapUIValue = { guibg = colors.dark_background },
+-- DapUIFrameName = { guibg = colors.dark_background },
+
+-- Nvim-tree
+-- NvimTreeFolderIcon = { guifg = base0D },
+-- NvimTreeFolderName = { guifg = colors.white },
+-- NvimTreeOpenedFolderName = { guifg = colors.white },
+-- NvimTreeEmptyFolderName = { guifg = colors.white },
+-- NvimTreeOpenedFile = { guifg = colors.white },
+-- NvimTreeIndentMarker = { guifg = colors.indent },
+-- NvimTreeVertSplit = { guifg = colors.dark_background, guibg = colors.dark_background },
+-- NvimTreeStatuslineNc = { guifg = colors.dark_background, guibg = colors.dark_background, gui = 'underline' },
+-- NvimTreeNormal = { guibg = colors.dark_background },
+-- NvimTreeRootFolder = { guifg = colors.white, gui = 'bold' },
 
 vim.g.terminal_color_0 = base00
 vim.g.terminal_color_1 = base08