+++ lua/setup/bufferline.lua
@@ -3,7 +3,10 @@ local sidebar = require("sidebar")
 
 require("bufferline").setup({
     options = {
-        indicator_icon = "▌",
+        indicator = {
+            icon = "▌",
+            stayle = "icon",
+        },
         diagnostics = "nvim_lsp",
         diagnostics_indicator = function(count, _)
             return "(" .. count .. ")"
@@ -21,130 +24,130 @@ require("bufferline").setup({
 
     highlights = {
         fill = {
-            guibg = colors.dark_background,
+            bg = colors.dark_background,
         },
         background = {
-            guibg = colors.dark_background,
+            bg = colors.dark_background,
         },
 
         tab = {
-            guifg = colors.comment,
-            guibg = colors.dark_background,
+            fg = colors.comment,
+            bg = colors.dark_background,
         },
         tab_selected = {
-            guifg = colors.white,
-            guibg = colors.background,
-            gui = "bold",
+            fg = colors.white,
+            bg = colors.background,
+            bold = true,
         },
 
         buffer_visible = {
-            guifg = colors.comment,
-            guibg = colors.dark_background,
+            fg = colors.comment,
+            bg = colors.dark_background,
         },
         buffer_selected = {
-            guifg = colors.white,
-            guibg = colors.background,
-            gui = "bold",
+            fg = colors.white,
+            bg = colors.background,
+            bold = true,
         },
 
         modified = {
-            guifg = colors.green,
-            guibg = colors.background,
+            fg = colors.green,
+            bg = colors.background,
         },
         modified_visible = {
-            guifg = colors.green,
-            guibg = colors.dark_background,
+            fg = colors.green,
+            bg = colors.dark_background,
         },
         modified_selected = {
-            guifg = colors.green,
-            guibg = colors.background,
+            fg = colors.green,
+            bg = colors.background,
         },
 
         error = {
-            guifg = colors.red,
-            guibg = colors.background,
+            fg = colors.red,
+            bg = colors.background,
         },
         error_visible = {
-            guifg = colors.red,
-            guibg = colors.dark_background,
+            fg = colors.red,
+            bg = colors.dark_background,
         },
         error_selected = {
-            guifg = colors.red,
-            guibg = colors.background,
+            fg = colors.red,
+            bg = colors.background,
         },
 
         error_diagnostic = {
-            guifg = colors.red,
-            guibg = colors.background,
+            fg = colors.red,
+            bg = colors.background,
         },
         error_diagnostic_visible = {
-            guifg = colors.red,
-            guibg = colors.dark_background,
+            fg = colors.red,
+            bg = colors.dark_background,
         },
         error_diagnostic_selected = {
-            guifg = colors.red,
-            guibg = colors.background,
+            fg = colors.red,
+            bg = colors.background,
         },
 
         warning = {
-            guifg = colors.yellow,
-            guibg = colors.background,
+            fg = colors.yellow,
+            bg = colors.background,
         },
         warning_visible = {
-            guifg = colors.yellow,
-            guibg = colors.dark_background,
+            fg = colors.yellow,
+            bg = colors.dark_background,
         },
         warning_selected = {
-            guifg = colors.yellow,
-            guibg = colors.background,
+            fg = colors.yellow,
+            bg = colors.background,
         },
 
         warning_diagnostic = {
-            guifg = colors.yellow,
-            guibg = colors.background,
+            fg = colors.yellow,
+            bg = colors.background,
         },
         warning_diagnostic_visible = {
-            guifg = colors.yellow,
-            guibg = colors.dark_background,
+            fg = colors.yellow,
+            bg = colors.dark_background,
         },
         warning_diagnostic_selected = {
-            guifg = colors.yellow,
-            guibg = colors.background,
+            fg = colors.yellow,
+            bg = colors.background,
         },
 
         info = {
-            guifg = colors.green,
-            guibg = colors.background,
+            fg = colors.green,
+            bg = colors.background,
         },
         info_visible = {
-            guifg = colors.green,
-            guibg = colors.dark_background,
+            fg = colors.green,
+            bg = colors.dark_background,
         },
         info_selected = {
-            guifg = colors.green,
-            guibg = colors.background,
+            fg = colors.green,
+            bg = colors.background,
         },
 
         info_diagnostic = {
-            guifg = colors.green,
-            guibg = colors.background,
+            fg = colors.green,
+            bg = colors.background,
         },
         info_diagnostic_visible = {
-            guifg = colors.green,
-            guibg = colors.dark_background,
+            fg = colors.green,
+            bg = colors.dark_background,
         },
         info_diagnostic_selected = {
-            guifg = colors.green,
-            guibg = colors.background,
+            fg = colors.green,
+            bg = colors.background,
         },
 
         indicator_selected = {
-            guifg = colors.blue,
-            guibg = colors.background,
+            fg = colors.blue,
+            bg = colors.background,
         },
 
         separator = {
-            guifg = colors.white,
+            fg = colors.white,
         },
     },
 })